chore: formatDate 컴포넌트 분리
chore: #67
