+++ src/api/formatDate.js
@@ -1,0 +1,9 @@
+const formatDate = (timestamp) => {
+  const date = new Date(timestamp);
+  const year = String(date.getFullYear()).slice(-2);
+  const month = String(date.getMonth() + 1).padStart(2, '0');
+  const day = String(date.getDate()).padStart(2, '0');
+  return `${year}-${month}-${day}`;
+};
+
+export default formatDate;
