+++ src/pages/Community/CommunityPostPage.jsx
@@ -3,6 +3,7 @@ import { useNavigate, useLocation } from 'react-router-dom';
 import styled from 'styled-components';
 import CommunityComment from '../../store/community/CommunityComment';
 import ImageModal from '../../components/Comunity/ImageModal';
+import formatDate from '../../api/formatDate';
 
 import ViewPhoto from '../../assets/Community/ViewPhoto.svg';
 import CommentPhoto from '../../assets/Community/CommentPhoto.svg';
@@ -17,14 +18,6 @@ const CommunityPostPage = () => {
 
   const [isModalOpen, setIsModalOpen] = useState(false);
   const [imgUrl, setImgUrl] = useState('');
-
-  const formatDate = (timestamp) => {
-    const date = new Date(timestamp);
-    const year = String(date.getFullYear()).slice(-2);
-    const month = String(date.getMonth() + 1).padStart(2, '0');
-    const day = String(date.getDate()).padStart(2, '0');
-    return `${year}-${month}-${day}`;
-  };
 
   const handleModal = (imgUrl) => {
     setImgUrl(imgUrl);
